backend/requirements.txt
Flask
openai
python-dotenv
pysrt
yt-dlp
flask-cors
httpx